+++ database/migrations/2025_05_05_143037_add_role_to_users_table.php
@@ -12,14 +12,20 @@ return new class extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->string('role')->default('user');  // Menambahkan kolom role dengan default 'user'
+            // Periksa apakah kolom role sudah ada
+            if (!Schema::hasColumn('users', 'role')) {
+                $table->string('role')->default('user')->nullable();  // Menambahkan kolom role dengan default 'user'
+            }
         });
     }
 
     public function down()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('role');  // Menghapus kolom role jika migrasi dibatalkan
+            // Hapus kolom role hanya jika ada
+            if (Schema::hasColumn('users', 'role')) {
+                $table->dropColumn('role');  // Menghapus kolom role jika migrasi dibatalkan
+            }
         });
     }
